backend/requirements.txt
absl-py==0.11.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
bcrypt==3.1.7
bleach==3.2.1
blinker==1.4
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.2.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dnspython==1.16.0
email-validator==1.1.1
entrypoints==0.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
flask-expects-json==1.5.0
Flask-JWT==0.3.2
Flask-Limiter==1.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
form==0.0.1
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.1
google-colab==1.0.0
google-pasta==0.2.0
grpcio==1.33.1
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
importlib-metadata==2.0.0
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
jsonify==0.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
jwt==1.1.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
limits==1.5.1
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.9.1
matplotlib==3.3.1
mistune==0.8.4
more-itertools==8.6.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
nltk==3.5
notebook==5.2.2
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==0.24.2
pandocfilters==1.4.3
passlib==1.7.4
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==7.2.0
portpicker==1.2.0
prompt-toolkit==1.0.18
protobuf==3.13.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.2
PyJWT==1.4.2
pyOpenSSL==20.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.4
pywin32==228
PyYAML==5.3.1
pyzmq==19.0.2
regex==2020.5.14
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.11.0
simplegeneric==0.8.1
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.18
svgwrite==1.4
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
testpath==0.4.4
textblob==0.15.3
threadpoolctl==2.1.0
tornado==4.5.3
tqdm==4.46.1
traitlets==5.0.5
Tree==0.2.4
tweepy==3.8.0
urllib3==1.24.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
zipp==3.4.0
